synthetic_data/requirements.txt
openai-whisper==20231117
torch==2.5.1
torchaudio==2.5.1
huggingface_hub
IPython
pyarrow
matplotlib
librosa
soundfile
datasets
encodec
boto3
fire
vector_quantize_pytorch==1.6.22
webdataset==0.2.85
whisperspeech
numpy==1.26.4
ray
vllm